refactor: attachment icons folder structure
