+++ web/components/icons/attachment/css-file-icon.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 // image
 import CssFileIcon from "public/attachment/css-icon.png";
 // type
-import type { ImageIconPros } from "./types";
+import type { ImageIconPros } from "../types";
 
 export const CssIcon: React.FC<ImageIconPros> = ({ width, height }) => (
   <Image src={CssFileIcon} height={height} width={width} alt="CssFileIcon" />
