+++ web/components/icons/attachment/document-icon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import type { Props } from "./types";
+import type { Props } from "../types";
 
 export const DocumentIcon: React.FC<Props> = ({ width = "24", height = "24", className, color }) => (
   <svg
