+++ web/components/icons/attachment/html-file-icon.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 // image
 import HtmlFileIcon from "public/attachment/html-icon.png";
 // type
-import type { ImageIconPros } from "./types";
+import type { ImageIconPros } from "../types";
 
 export const HtmlIcon: React.FC<ImageIconPros> = ({ width, height }) => (
   <Image src={HtmlFileIcon} height={height} width={width} alt="HtmlFileIcon" />
