+++ web/components/icons/attachment/index.ts
@@ -1,0 +1,20 @@
+export * from "./attachment-icon";
+export * from "./audio-file-icon";
+export * from "./css-file-icon";
+export * from "./csv-file-icon";
+export * from "./default-file-icon";
+export * from "./doc-file-icon";
+export * from "./document-icon";
+export * from "./figma-file-icon";
+export * from "./html-file-icon";
+export * from "./img-file-icon";
+export * from "./jpg-file-icon";
+export * from "./js-file-icon";
+export * from "./pdf-file-icon";
+export * from "./png-file-icon";
+export * from "./setting-icon";
+export * from "./sheet-file-icon";
+export * from "./svg-file-icon";
+export * from "./tune-icon";
+export * from "./txt-file-icon";
+export * from "./video-file-icon";
