+++ web/components/icons/attachment/svg-file-icon.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 // image
 import SvgFileIcon from "public/attachment/svg-icon.png";
 // type
-import type { ImageIconPros } from "./types";
+import type { ImageIconPros } from "../types";
 
 export const SvgIcon: React.FC<ImageIconPros> = ({ width, height }) => (
   <Image src={SvgFileIcon} height={height} width={width} alt="SvgFileIcon" />
