+++ web/components/icons/attachment/tune-icon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import type { Props } from "./types";
+import type { Props } from "../types";
 
 export const TuneIcon: React.FC<Props> = ({ width = "24", height = "24", className }) => (
   <svg
