+++ web/components/icons/attachment/video-file-icon.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 // image
 import VideoFileIcon from "public/attachment/video-icon.png";
 // type
-import type { ImageIconPros } from "./types";
+import type { ImageIconPros } from "../types";
 
 export const VideoIcon: React.FC<ImageIconPros> = ({ width, height }) => (
   <Image src={VideoFileIcon} height={height} width={width} alt="VideoFileIcon" />
